% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder.R
\name{encoder}
\alias{encoder}
\title{encoder - R6 Class for Categorical Data Encoding}
\description{
This R6 class provides methods for one-hot encoding and label encoding
categorical data.
}
\examples{
encoder_ <- encoder$new()

## ------------------------------------------------
## Method `encoder$OneHotEncode`
## ------------------------------------------------

data <- data.frame(category = c('A', 'B', 'A', 'C'))
encoder_ <- encoder$new()
encoded_data <- encoder_$OneHotEncode(data)

## ------------------------------------------------
## Method `encoder$LabelEncode`
## ------------------------------------------------

data <- data.frame(category = c('A', 'B', 'A', 'C'))
encoder_ <- encoder$new()
encoded_data <- encoder_$LabelEncode(data)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-encoder-OneHotEncode}{\code{encoder$OneHotEncode()}}
\item \href{#method-encoder-LabelEncode}{\code{encoder$LabelEncode()}}
\item \href{#method-encoder-clone}{\code{encoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encoder-OneHotEncode"></a>}}
\if{latex}{\out{\hypertarget{method-encoder-OneHotEncode}{}}}
\subsection{Method \code{OneHotEncode()}}{
This function performs one-hot encoding on categorical data,
converting categorical variables into binary vectors. Each unique
category becomes a binary column in the resulting data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{encoder$OneHotEncode(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame containing categorical variables to be
one-hot encoded.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with the original quantitative columns and
         additional columns representing the one-hot encoded categorical
         variables.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data <- data.frame(category = c('A', 'B', 'A', 'C'))
encoder_ <- encoder$new()
encoded_data <- encoder_$OneHotEncode(data)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encoder-LabelEncode"></a>}}
\if{latex}{\out{\hypertarget{method-encoder-LabelEncode}{}}}
\subsection{Method \code{LabelEncode()}}{
This function performs label encoding on categorical data, converting
categorical variables into numeric labels. Each unique category is
assigned a unique integer label in the resulting data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{encoder$LabelEncode(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame containing categorical variables to be label
encoded.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with the original columns replaced by numeric
     labels for the categorical variables.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data <- data.frame(category = c('A', 'B', 'A', 'C'))
encoder_ <- encoder$new()
encoded_data <- encoder_$LabelEncode(data)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-encoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{encoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
