% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{metrics}
\description{
This class implements methods to evaluate model performances
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-metrics-clone}{\code{metrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-metrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-metrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{metrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
