% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes.R
\name{naive_bayes}
\alias{naive_bayes}
\title{naive_bayes - R6 Class for Naive Bayes Classification}
\description{
This class implements a supervised learning methods based on applying Bayes’
theorem with strong (naive) feature independence assumption
}
\examples{

## ------------------------------------------------
## Method `naive_bayes$fit`
## ------------------------------------------------

data <- data.frame(
  feature1 = c(1, 2, 3, 4),
  feature2 = c(5, 6, 7, 8),
  class = c('A', 'B', 'A', 'B')
)
classifier <- naive_bayes$new()
classifier$fit(data[, c('feature1', 'feature2')], data$class)


## ------------------------------------------------
## Method `naive_bayes$predict`
## ------------------------------------------------

new_data <- data.frame(feature1 = c(1, 2), feature2 = c(5, 6))
ypred <- classifier$predict(new_data)
print(ypred)


## ------------------------------------------------
## Method `naive_bayes$predict_proba`
## ------------------------------------------------

new_data <- data.frame(feature1 = c(1, 2), feature2 = c(5, 6))
probabilities <- classifier$predict_proba(new_data)
print(probabilities)


## ------------------------------------------------
## Method `naive_bayes$print`
## ------------------------------------------------

print(classifier)


## ------------------------------------------------
## Method `naive_bayes$summary`
## ------------------------------------------------

classifier$summary()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classes_}}{distinct classes}

\item{\code{theta_}}{Mean of each feature per class}

\item{\code{sigma_}}{Variance of each feature per class}

\item{\code{prior_}}{probability of each class}

\item{\code{feature_names_in_}}{Names of features seen during fit}

\item{\code{feature_dtypes_}}{types of features seen during fit}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-naive_bayes-new}{\code{naive_bayes$new()}}
\item \href{#method-naive_bayes-fit}{\code{naive_bayes$fit()}}
\item \href{#method-naive_bayes-predict}{\code{naive_bayes$predict()}}
\item \href{#method-naive_bayes-predict_proba}{\code{naive_bayes$predict_proba()}}
\item \href{#method-naive_bayes-print}{\code{naive_bayes$print()}}
\item \href{#method-naive_bayes-summary}{\code{naive_bayes$summary()}}
\item \href{#method-naive_bayes-clone}{\code{naive_bayes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-new"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$new(multi_thread = FALSE, n_cluster = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{multi_thread}}{Logical, indicating whether fit method will be
parallelize. Default is FALSE.}

\item{\code{n_cluster}}{Numeric, indicating the number of cpu to use for
parallel computing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-fit"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-fit}{}}}
\subsection{Method \code{fit()}}{
This method trains a Naive Bayes classifier on the given training data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$fit(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A data frame or matrix containing the feature variables.}

\item{\code{y}}{A vector or factor containing the class labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the trained naive_bayes object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data <- data.frame(
  feature1 = c(1, 2, 3, 4),
  feature2 = c(5, 6, 7, 8),
  class = c('A', 'B', 'A', 'B')
)
classifier <- naive_bayes$new()
classifier$fit(data[, c('feature1', 'feature2')], data$class)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-predict"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-predict}{}}}
\subsection{Method \code{predict()}}{
This method predicts class labels for new data using the trained Naive
Bayes classifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$predict(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A data frame or matrix containing the feature variables
for which to predict class labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predicted class labels.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{new_data <- data.frame(feature1 = c(1, 2), feature2 = c(5, 6))
ypred <- classifier$predict(new_data)
print(ypred)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
This method predicts class probabilities for new data using the trained
Naive Bayes classifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$predict_proba(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A data frame or matrix containing the feature variables
for which to predict class probabilities.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with columns representing class labels and their
        corresponding probabilities for each observation in new_data.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{new_data <- data.frame(feature1 = c(1, 2), feature2 = c(5, 6))
probabilities <- classifier$predict_proba(new_data)
print(probabilities)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-print"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-print}{}}}
\subsection{Method \code{print()}}{
This method prints the naive_bayes object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters to be passed to the print method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Prints of the naive_bayes object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{print(classifier)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-summary"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-summary}{}}}
\subsection{Method \code{summary()}}{
This method prints a summary of the naive_bayes object,
including information about the trained model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prints a summary of the naive_bayes object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{classifier$summary()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
