% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoder.R
\name{one_hot_encoder}
\alias{one_hot_encoder}
\title{one_hot_encoder - R6 Class for Categorical Data Encoding.}
\description{
This R6 class provides methods for one-hot encoding categorical data.
}
\examples{
encoder_ <- one_hot_encoder$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-one_hot_encoder-fit}{\code{one_hot_encoder$fit()}}
\item \href{#method-one_hot_encoder-transform}{\code{one_hot_encoder$transform()}}
\item \href{#method-one_hot_encoder-is_fitted}{\code{one_hot_encoder$is_fitted()}}
\item \href{#method-one_hot_encoder-clone}{\code{one_hot_encoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-one_hot_encoder-fit"></a>}}
\if{latex}{\out{\hypertarget{method-one_hot_encoder-fit}{}}}
\subsection{Method \code{fit()}}{
Method to train the encoder.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{one_hot_encoder$fit(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to be used for training.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The trained OneHotEncoder object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-one_hot_encoder-transform"></a>}}
\if{latex}{\out{\hypertarget{method-one_hot_encoder-transform}{}}}
\subsection{Method \code{transform()}}{
Method to perform one-hot encoding of data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{one_hot_encoder$transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to be transformed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The encoded data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-one_hot_encoder-is_fitted"></a>}}
\if{latex}{\out{\hypertarget{method-one_hot_encoder-is_fitted}{}}}
\subsection{Method \code{is_fitted()}}{
Method to check if the encoder is trained.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{one_hot_encoder$is_fitted()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if the encoder is trained, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-one_hot_encoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-one_hot_encoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{one_hot_encoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
