% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_indices_stratified.R
\name{split_indices_stratified}
\alias{split_indices_stratified}
\title{Title}
\usage{
split_indices_stratified(labels, train_size)
}
\arguments{
\item{labels}{A vector or factor containing the class labels}

\item{train_size}{The proportion of the dataset to include in the
train set. Should be a value between 0 and 1.}
}
\value{
A list containing the indexes of train samples
}
\description{
Title
}
