% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{train_test_split}
\usage{
train_test_split(data, train_size = 0.7, stratify = NULL, seed = NULL)
}
\arguments{
\item{data}{A data frame or matrix containing the dataset to be split.}

\item{train_size}{The proportion of the dataset to include in the
train set. Should be a value between 0 and 1.}

\item{stratify}{A character indicating the name of class labels for
stratified sampling. If provided, the split will
maintain the same distribution of class labels in
both the train and test sets.}

\item{seed}{An optional seed for reproducibility. If NULL, no seed is set.}
}
\value{
A list containing the train and test sets.
}
\description{
This function splits the input data into train and test sets.
}
\examples{
data <- data.frame(
  features = rnorm(100),
  labels = sample(c('A', 'B'), 100, replace = TRUE)
)
split_data <- train_test_split(
  data,
  train_size = 0.7,
  stratify = "labels",
  seed = 123
)
print(head(split_data$train_set))
print(head(split_data$test_set))

}
